FIX - Adição de Despesas e Receitas
Correção do layout das telas de adição de receitas e despesas
